+++ packages/react/src/index.tsx
@@ -6,6 +6,8 @@ export * from './components/Checkbox'
 export * from './components/Badge'
 export * from './components/Status'
 export * from './components/CardInfoCheck'
+export * from './components/CardCountCheck'
+export * from './components/CardNotification'
 export * from './components/Select'
 
 export * from './styles'