+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @peoplecheck-ui/react
 
+## 2.1.0
+
+### Minor Changes
+
+- Insert new components
+
 ## 2.0.0
 
 ### Major Changes